projects/ml (1)/ml-models/src/docai-models-stayconnected/requirements.txt
numpy==1.20.0
python-Levenshtein==0.12.2
fuzzywuzzy==0.18.0
Unidecode==1.3.3
unicodedata2==13.0.0.post2
opencv-python-headless==4.5.5.62
google-cloud-aiplatform==1.12.0
google-cloud==0.34.0
Pillow==9.0.0